* arts-2-0-23
* CMakeLists.txt:

Add an additional check right at the beginning to make sure that we
never allow cmake to be called inside the source directory.

NOTE: If you get an error after updating that you have to build
arts outside the source tree although you are actually in a separate
build directory, make sure that there is no file 'CMakeCache.txt' or
directory 'CMakeFiles' in your toplevel arts source(!!) directory.

* tests/testall.py: Fix path to arts for Xcode build.

* src/docserver.cc (insert_breadcrumbs): Fix type of t, int -> size_t.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7220 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/docserver.cc
@@ -1526,7 +1526,7 @@ void Docserver::insert_breadcrumb_token (size_t token_id)
 void Docserver::insert_breadcrumbs ()
 {
   get_os() << "<div id=\"navbar\"><div class=\"breadcrumbs\">";
-  for (int t = 0; t <= tokens.size(); t++)
+  for (size_t t = 0; t <= tokens.size(); t++)
   {
     if (t) get_os() << "&nbsp;>>&nbsp;";
     insert_breadcrumb_token(t);
